Rafactor: Made code reuseable

+++ internunity/app/Http/Controllers/PortfolioController.php
@@ -5,23 +5,27 @@ namespace App\Http\Controllers;
 use App\Http\Requests\PortfolioCreateRequest;
 use App\Models\Project;
 use App\Services\File;
-use Illuminate\Http\Request;
-use Laravel\Sanctum\PersonalAccessToken;
 
 class PortfolioController extends Controller
 {
     private $per_page = 15;
-    public function get(Request $request) {
-        $token = PersonalAccessToken::findToken($request->bearerToken());
-        return $token->tokenable->projects()->orderBy("created_at", "DESC")->paginate($this->per_page);
+    protected $user;
+    protected $user_service;
+    public function __construct(\App\Models\User $user) {
+        $this->user = $user->getUser();
+        $this->user_service = $user;
+    }
+    public function get(): mixed {
+        return $this->user->projects()->orderBy("created_at", "DESC")->paginate($this->per_page);
     }
     public function store(PortfolioCreateRequest $request, File $file) {
-        $token = PersonalAccessToken::findToken($request->bearerToken());
         $fileName = $file->upload($request->file("project_file"));
-        $token->tokenable->projects()->create([ ...$request->validated(), "to" => now(), "from" => now(), "project_pic" => $fileName ]);
+        $this->user->projects()->create([ ...$request->validated(), "to" => now(), "from" => now(), "project_pic" => $fileName ]);
         return 1;
     }
     public function edit(PortfolioCreateRequest $request, Project $project, File $file) {
+        if(!$this->user_service->isOwnerOf($project)) abort(403);
+
         $fileName = null;
         if($request->hasFile("project_file"))
             $fileName = $file->replace($project->project_pic, $request->file("project_file"));
@@ -29,6 +33,7 @@ class PortfolioController extends Controller
         return 1;
     }
     public function destroy(Project $project) {
+        if(!$this->user_service->isOwnerOf($project)) abort(403);
         $project->delete();
         return 1;
     }
